feat: mainPage의 imagesection 목데이터 실제데이터로 변경
