+++ src/components/product/main/cards.ts
@@ -1,0 +1,27 @@
+export const cards = [
+  {
+    id: 'card1',
+    title: '1번설명',
+    image: '/image/exp1.png',
+  },
+  {
+    id: 'card2',
+    title: '2번설명',
+    image: '/image/exp2.png',
+  },
+  {
+    id: 'card3',
+    title: '3번설명',
+    image: '/image/exp3.png',
+  },
+  {
+    id: 'card4',
+    title: '4번설명',
+    image: '/image/exp4.png',
+  },
+  {
+    id: 'card5',
+    title: '5번설명',
+    image: '/image/exp5.png',
+  },
+];
