--- src/mock/cards.ts
@@ -1,42 +1,0 @@
-export const cards = [
-  {
-    id: 'card1',
-    title: 'Blonkisoaz1',
-    subtitle: 'Omuke trughte a otufta1',
-    icon: 'fas fa-walking',
-    image:
-      'https://66.media.tumblr.com/6fb397d822f4f9f4596dff2085b18f2e/tumblr_nzsvb4p6xS1qho82wo1_1280.jpg',
-  },
-  {
-    id: 'card2',
-    title: 'Blonkisoaz2',
-    subtitle: 'Omuke trughte a otufta2',
-    icon: 'fas fa-walking',
-    image:
-      'https://66.media.tumblr.com/6fb397d822f4f9f4596dff2085b18f2e/tumblr_nzsvb4p6xS1qho82wo1_1280.jpg',
-  },
-  {
-    id: 'card3',
-    title: 'Blonkisoaz3',
-    subtitle: 'Omuke trughte a otufta3',
-    icon: 'fas fa-walking',
-    image:
-      'https://66.media.tumblr.com/6fb397d822f4f9f4596dff2085b18f2e/tumblr_nzsvb4p6xS1qho82wo1_1280.jpg',
-  },
-  {
-    id: 'card4',
-    title: 'Blonkisoaz4',
-    subtitle: 'Omuke trughte a otufta4',
-    icon: 'fas fa-walking',
-    image:
-      'https://66.media.tumblr.com/6fb397d822f4f9f4596dff2085b18f2e/tumblr_nzsvb4p6xS1qho82wo1_1280.jpg',
-  },
-  {
-    id: 'card5',
-    title: 'Blonkisoaz5',
-    subtitle: 'Omuke trughte a otufta5',
-    icon: 'fas fa-walking',
-    image:
-      'https://66.media.tumblr.com/6fb397d822f4f9f4596dff2085b18f2e/tumblr_nzsvb4p6xS1qho82wo1_1280.jpg',
-  },
-];
